+++ src/index.js
@@ -6,7 +6,13 @@ const {
 const { toGrowthHubHint } = require("./adapters/growth_hub/adapter");
 const { validateHabitRules } = require("./habit_registry/validate_registry");
 const {
+  PACKAGE_NAME,
+  USER_HOME_OVERRIDE_ENV
+} = require("./runtime_paths");
+const {
+  LEGACY_USER_REGISTRY_PATH,
   USER_REGISTRY_PATH,
+  USER_DATA_ROOT,
   addUserHabitRule,
   ensureUserRegistryFile,
   exportUserRegistryState,
@@ -14,6 +20,8 @@ const {
   loadMergedHabits,
   loadUserRegistryState,
   removeUserHabitPhrase,
+  resolveDefaultUserRegistryPath,
+  resolveUserDataRoot,
   suppressSuggestionPhrase
 } = require("./habit_registry/user_registry");
 const { parseHabitManagementRequest } = require("./habit_registry/management_prompt");
@@ -31,7 +39,11 @@ const {
 } = require("./session_suggestions/cache");
 
 module.exports = {
+  LEGACY_USER_REGISTRY_PATH,
+  PACKAGE_NAME,
   USER_REGISTRY_PATH,
+  USER_DATA_ROOT,
+  USER_HOME_OVERRIDE_ENV,
   addUserHabitRule,
   ensureUserRegistryFile,
   exportUserRegistryState,
@@ -49,6 +61,8 @@ module.exports = {
   parseCandidateReference,
   parseSessionTranscript,
   removeUserHabitPhrase,
+  resolveDefaultUserRegistryPath,
+  resolveUserDataRoot,
   suppressSuggestionPhrase,
   saveSuggestionSnapshot,
   suggestSessionHabitCandidates,